pplbackupbot/requirements.txt
flask_cors
# openai
tiktoken